pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-doc-processing"
version = "0.1.0"
description = "Agentic document processing with LangChain agent: assessment, layout (Surya), OCR/HTR, export."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "google-genai>=1.0.0",
    "surya-ocr>=0.17.0",
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "matplotlib>=3.7.0",
    "numpy>=1.21.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]
tesseract = ["pytesseract"]
trocr = ["torch", "transformers"]

[project.scripts]
agentic-doc = "main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
